+++ lib/extensions/extensions.dart
@@ -1,0 +1,4 @@
+library extensions;
+
+export 'list_extension.dart';
+export 'string_extension.dart';