+++ lib/widgets/widgets.dart
@@ -1,0 +1,3 @@
+library widgets;
+
+export 'input_box.dart';