+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+* Added an intersect method on lists
+* Added a widget library
+* Added a extensions library
+
 ## 0.0.2
 * Added an input box widget